bugprone-narrowing-conversions

+++ src/fastq_stats.cc
@@ -342,7 +342,7 @@ auto report_read_length_distribution(std::FILE * fp_log, std::vector<uint64_t> c
                        (i == len_max ? ">=" : "  "),
                        i,
                        read_length_table[i],
-                       read_length_table[i] * 100.0 / seq_count,
+                       static_cast<double>(read_length_table[i]) * 100.0 / seq_count,
                        100.0 * (seq_count - (i > 0 ? length_dist[i - 1] : 0)) / seq_count);
         }
       if (i == 0UL) { break; }
